+++ src/azure-cli/azure/cli/command_modules/netappfiles/tests/latest/test_volume_commands.py
@@ -7,7 +7,7 @@ from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
 
 POOL_DEFAULT = "--service-level 'Premium' --size 4"
 VOLUME_DEFAULT = "--service-level 'Premium' --usage-threshold 100"
-LOCATION = "westcentralus"
+LOCATION = "westus2"
 GIB_SCALE = 1024 * 1024 * 1024
 
 # No tidy up of tests required. The resource group is automatically removed
@@ -24,19 +24,19 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
 
     def create_volume(self, account_name, pool_name, volume_name1, rg, tags=None, volume_name2=None, protocols=None, pool_payload=POOL_DEFAULT, volume_payload=VOLUME_DEFAULT):
         vnet_name = self.create_random_name(prefix='cli-vnet-', length=24)
-        creation_token = volume_name1
+        file_path = volume_name1  # creation_token
         subnet_name = self.create_random_name(prefix='cli-subnet-', length=16)
         tag = "--tags %s" % tags if tags is not None else ""
         protocol_types = "--protocol-types %s" % protocols if protocols is not None else ""
 
-        self.setup_vnet(rg, vnet_name, subnet_name, '10.12.0.0')
+        self.setup_vnet(rg, vnet_name, subnet_name, '10.0.0.0')
         self.cmd("az netappfiles account create -g %s -a '%s' -l %s" % (rg, account_name, LOCATION)).get_output_in_json()
         self.cmd("az netappfiles pool create -g %s -a %s -p %s -l %s %s %s" % (rg, account_name, pool_name, LOCATION, pool_payload, tag)).get_output_in_json()
-        volume1 = self.cmd("az netappfiles volume create --resource-group %s --account-name %s --pool-name %s --volume-name %s -l %s %s --creation-token %s --vnet %s --subnet %s %s %s" % (rg, account_name, pool_name, volume_name1, LOCATION, volume_payload, creation_token, vnet_name, subnet_name, protocol_types, tag)).get_output_in_json()
+        volume1 = self.cmd("az netappfiles volume create --resource-group %s --account-name %s --pool-name %s --volume-name %s -l %s %s --file-path %s --vnet %s --subnet %s %s %s" % (rg, account_name, pool_name, volume_name1, LOCATION, volume_payload, file_path, vnet_name, subnet_name, protocol_types, tag)).get_output_in_json()
 
         if volume_name2:
-            creation_token = volume_name2
-            self.cmd("az netappfiles volume create -g %s -a %s -p %s -v %s -l %s %s --creation-token %s --vnet %s --subnet %s --tags %s" % (rg, account_name, pool_name, volume_name2, LOCATION, VOLUME_DEFAULT, creation_token, vnet_name, subnet_name, tags)).get_output_in_json()
+            file_path = volume_name2
+            self.cmd("az netappfiles volume create -g %s -a %s -p %s -v %s -l %s %s --file-path %s --vnet %s --subnet %s --tags %s" % (rg, account_name, pool_name, volume_name2, LOCATION, VOLUME_DEFAULT, file_path, vnet_name, subnet_name, tags)).get_output_in_json()
 
         return volume1
 
@@ -47,8 +47,7 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
         volume_name = self.create_random_name(prefix='cli-vol-', length=24)
         tags = "Tag1=Value1 Tag2=Value2"
 
-        protocol_types = "NFSv3 NFSv4"
-
+        protocol_types = "NFSv3 NFSv4.1"
         volume = self.create_volume(account_name, pool_name, volume_name, '{rg}', tags=tags, protocols=protocol_types)
         assert volume['name'] == account_name + '/' + pool_name + '/' + volume_name
         assert volume['tags']['Tag1'] == 'Value1'
@@ -62,7 +61,7 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
         # specified protocol type
         assert len(volume['protocolTypes']) == 2
         assert volume['protocolTypes'][0] == 'NFSv3'
-        assert volume['protocolTypes'][1] == 'NFSv4'
+        assert volume['protocolTypes'][1] == 'NFSv4.1'
 
         volume_list = self.cmd("netappfiles volume list --resource-group {rg} --account-name %s --pool-name %s" % (account_name, pool_name)).get_output_in_json()
         assert len(volume_list) == 1
@@ -70,6 +69,33 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
         self.cmd("az netappfiles volume delete --resource-group {rg} --account-name %s --pool-name %s --volume-name %s" % (account_name, pool_name, volume_name))
         volume_list = self.cmd("netappfiles volume list --resource-group {rg} -a %s -p %s" % (account_name, pool_name)).get_output_in_json()
         assert len(volume_list) == 0
+
+    @ResourceGroupPreparer(name_prefix='cli_tests_rg')
+    def test_create_volume_with_subnet_in_different_rg(self):
+        account_name = self.create_random_name(prefix='cli-acc-', length=24)
+        pool_name = self.create_random_name(prefix='cli-pool-', length=24)
+        volume_name = self.create_random_name(prefix='cli-vol-', length=24)
+
+        vnet_name = self.create_random_name(prefix='cli-vnet-', length=24)
+        file_path = volume_name  # creation_token
+        subnet_name = self.create_random_name(prefix='cli-subnet-', length=16)
+
+        subnet_rg = self.create_random_name(prefix='cli-rg-', length=24)
+        subs_id = self.current_subscription()
+        self.cmd("az group create -n %s --subscription %s -l %s" % (subnet_rg, subs_id, LOCATION)).get_output_in_json()
+
+        rg = '{rg}'
+        self.setup_vnet(subnet_rg, vnet_name, subnet_name, '10.0.0.0')
+        self.cmd("az netappfiles account create -g %s -a %s -l %s" % (rg, account_name, LOCATION)).get_output_in_json()
+        self.cmd("az netappfiles pool create -g %s -a %s -p %s -l %s %s" % (rg, account_name, pool_name, LOCATION, POOL_DEFAULT)).get_output_in_json()
+
+        subnet_id = "/subscriptions/%s/resourceGroups/%s/providers/Microsoft.Network/virtualNetworks/%s/subnets/%s" % (subs_id, subnet_rg, vnet_name, subnet_name)
+
+        volume = self.cmd("az netappfiles volume create --resource-group %s --account-name %s --pool-name %s --volume-name %s -l %s %s --file-path %s --vnet %s --subnet %s" % (rg, account_name, pool_name, volume_name, LOCATION, VOLUME_DEFAULT, file_path, vnet_name, subnet_id)).get_output_in_json()
+        assert volume['name'] == account_name + '/' + pool_name + '/' + volume_name
+
+        self.cmd("az netappfiles volume delete --resource-group %s --account-name %s --pool-name %s --volume-name %s" % (rg, account_name, pool_name, volume_name))
+        self.cmd("az group delete --yes -n %s" % (subnet_rg))
 
     @ResourceGroupPreparer(name_prefix='cli_tests_rg')
     def test_list_volumes(self):
@@ -95,12 +121,16 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
         volume_name = self.create_random_name(prefix='cli-vol-', length=24)
         tags = "Tag2=Value1"
 
-        protocol_types = "NFSv4"
+        protocol_types = "NFSv4.1"
         volume = self.create_volume(account_name, pool_name, volume_name, '{rg}', tags=tags, protocols=protocol_types)
         assert volume['name'] == account_name + '/' + pool_name + '/' + volume_name
         # specified protocol type
         assert len(volume['protocolTypes']) == 1
-        assert volume['protocolTypes'][0] == 'NFSv4'
+        assert volume['protocolTypes'][0] == 'NFSv4.1'
+        assert len(volume['exportPolicy']['rules']) == 1
+        assert volume['exportPolicy']['rules'][0]['ruleIndex'] == 1
+        assert volume['exportPolicy']['rules'][0]['nfsv41']
+        assert not volume['exportPolicy']['rules'][0]['nfsv3']
 
         volume = self.cmd("az netappfiles volume show --resource-group {rg} -a %s -p %s -v %s" % (account_name, pool_name, volume_name)).get_output_in_json()
         assert volume['name'] == account_name + '/' + pool_name + '/' + volume_name
@@ -121,10 +151,12 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
         # default protocol type
         assert len(volume['protocolTypes']) == 1
         assert volume['protocolTypes'][0] == 'NFSv3'
+        assert volume['usageThreshold'] == 100 * GIB_SCALE
 
-        volume = self.cmd("az netappfiles volume update --resource-group {rg} -a %s -p %s -v %s --tags %s --service-level 'Standard'" % (account_name, pool_name, volume_name, tags)).get_output_in_json()
+        volume = self.cmd("az netappfiles volume update --resource-group {rg} -a %s -p %s -v %s --tags %s --usage-threshold 200" % (account_name, pool_name, volume_name, tags)).get_output_in_json()
         assert volume['name'] == account_name + '/' + pool_name + '/' + volume_name
-        assert volume['serviceLevel'] == "Standard"
+        assert volume['serviceLevel'] == "Premium"  # unchanged
+        assert volume['usageThreshold'] == 200 * GIB_SCALE
         assert volume['tags']['Tag1'] == 'Value2'
         # default export policy still present
         assert volume['exportPolicy']['rules'][0]['allowedClients'] == '0.0.0.0/0'
@@ -142,14 +174,14 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
 
         # now add an export policy
         # there is already one default rule present
-        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.3.0/24' --rule-index 3 --unix-read-only true --unix-read-write false --cifs false --nfsv3 true --nfsv4 false" % (account_name, pool_name, volume_name)).get_output_in_json()
+        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.3.0/24' --rule-index 3 --unix-read-only true --unix-read-write false --cifs false --nfsv3 true --nfsv41 false" % (account_name, pool_name, volume_name)).get_output_in_json()
         assert vol_with_export_policy['name'] == account_name + '/' + pool_name + '/' + volume_name
         assert vol_with_export_policy['exportPolicy']['rules'][0]['allowedClients'] == '1.2.3.0/24'
         assert vol_with_export_policy['exportPolicy']['rules'][0]['ruleIndex'] == 3
         assert vol_with_export_policy['exportPolicy']['rules'][0]['cifs'] is False
 
         # and add another export policy
-        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.4.0/24' --rule-index 2 --unix-read-only true --unix-read-write false --cifs true --nfsv3 true --nfsv4 false" % (account_name, pool_name, volume_name)).get_output_in_json()
+        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.4.0/24' --rule-index 2 --unix-read-only true --unix-read-write false --cifs true --nfsv3 true --nfsv41 false" % (account_name, pool_name, volume_name)).get_output_in_json()
         assert vol_with_export_policy['name'] == account_name + '/' + pool_name + '/' + volume_name
         assert vol_with_export_policy['exportPolicy']['rules'][1]['allowedClients'] == '1.2.3.0/24'
         assert vol_with_export_policy['exportPolicy']['rules'][0]['allowedClients'] == '1.2.4.0/24'
@@ -182,7 +214,7 @@ class AzureNetAppFilesVolumeServiceScenarioTest(ScenarioTest):
 
         # now add an export policy
         # there is already one default rule present
-        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.3.0/24' --rule-index 3 --unix-read-only true --unix-read-write false --cifs false --nfsv3 true --nfsv4 false" % (account_name, pool_name, volume_name)).get_output_in_json()
+        vol_with_export_policy = self.cmd("netappfiles volume export-policy add -g {rg} -a %s -p %s -v %s --allowed-clients '1.2.3.0/24' --rule-index 3 --unix-read-only true --unix-read-write false --cifs false --nfsv3 true --nfsv41 false" % (account_name, pool_name, volume_name)).get_output_in_json()
         assert vol_with_export_policy['name'] == account_name + '/' + pool_name + '/' + volume_name
         assert vol_with_export_policy['exportPolicy']['rules'][0]['allowedClients'] == '1.2.3.0/24'
         assert vol_with_export_policy['exportPolicy']['rules'][0]['ruleIndex'] == 3